% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colClass_as.R
\name{colClass_as}
\alias{colClass_as}
\title{colClass_as}
\usage{
colClass_as(data, type)
}
\arguments{
\item{data}{data.frame only. The dataset to be modified}

\item{type}{Vector of strings. Only four class are supported c("logical","numeric","factor","character")}
}
\value{
data.frame.
}
\description{
colClass_as
}
\examples{
if (interactive()) {
# Launch with built-in PK data set:
colClass_as(PK,c("character","numeric","factor","character"))
}

}
