% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiplot.R
\name{guiplot}
\alias{guiplot}
\title{guiplot}
\usage{
guiplot(..., out_dir = NULL)
}
\arguments{
\item{...}{Matrix or data frame}

\item{out_dir}{The storage path of the output picture, recommend 'out_dir=getwb()'}
}
\value{
png and pdf of plot
}
\description{
guiplot
}
\examples{
\donttest{
guiplot()
}
\donttest{
guiplot(PK)
}
\donttest{
guiplot(PK,out_dir=getwb())
}
}
