% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC_Class_Mapping_Table.R
\docType{class}
\name{Mapping_Table-class}
\alias{Mapping_Table-class}
\alias{Mapping_Table}
\title{A Reference Class to handle mapping_table events}
\description{
A Reference Class to handle mapping_table events
}
\section{Fields}{

\describe{
\item{\code{field_groups}}{matrix.}

\item{\code{variable}}{A character vector.}

\item{\code{default_field}}{A length-one numeric vector.}

\item{\code{preselection_field}}{A matrix.}

\item{\code{mapping_table}}{A matrix.}

\item{\code{field_index_transform}}{A matrix.}

\item{\code{mapping_table}}{A matrix.}

\item{\code{inf_of_mptbl}}{A matrix.}

\item{\code{field_right_bound}}{A vector}

\item{\code{DT_container}}{A character}
}}

\section{Methods}{

\describe{
\item{\code{assign_field_by_field_group()}}{assign field by field_group}

\item{\code{create_mptbl()}}{create a mapping_table, and select default field,and set field_right bound, set DT container}

\item{\code{get_sub_group(group_name)}}{Internal method: get sub_group from mapping_table}

\item{\code{mptbl_event_fill(fill_inf)}}{mapping_table fill event assign}

\item{\code{mptbl_event_select(selection_inf)}}{mapping_table select event assign}

\item{\code{mptbl_select_default_field(default_field)}}{mapping_table select default field}

\item{\code{mptbl_select_preselection_field(preselection_field)}}{mapping_table select preselection field}

\item{\code{mptbl_to_inf(mapping_table)}}{Internal method: Convert mapping_table to inf_of_mptbl}

\item{\code{set_DT_container(field_groups)}}{Internal method: set DT container}

\item{\code{set_field_right_bound(field_groups)}}{Internal method: set field right bound}

\item{\code{set_mptabl(group_name, mptbl)}}{Internal method: Use the value of the subtable to set the value of the}
}}

