% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC_Class_Mapping_Table.R
\docType{class}
\name{Single_Mptbl-class}
\alias{Single_Mptbl-class}
\alias{Single_Mptbl}
\title{A Reference Class to handle single-choice events}
\description{
A Reference Class to handle single-choice events
}
\section{Fields}{

\describe{
\item{\code{field}}{A length-one numeric vector.}

\item{\code{variable}}{A character vector.}

\item{\code{default_field}}{A length-one numeric vector.}

\item{\code{preselection_field}}{A matrix.}

\item{\code{mapping_table}}{A matrix.}
}}

\section{Methods}{

\describe{
\item{\code{create_mptbl()}}{create a mapping_table}

\item{\code{mptbl_event_fill(fill_inf)}}{Handle fill selection events}

\item{\code{mptbl_event_select(selection_inf)}}{Handle single-choice selection events}

\item{\code{mptbl_select_default_field(default_field)}}{mapping_table select default field}
}}

