% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiplot.R
\name{guiplot}
\alias{guiplot}
\title{guiplot}
\usage{
guiplot(..., out_dir = getwd())
}
\arguments{
\item{...}{Matrix or data frame}

\item{out_dir}{The storage path of the output picture, recommend 'out_dir=getwd()'}
}
\value{
Export files(png and pdf of plot) to a temporary directory, or user-defined folders.
}
\description{
guiplot
}
\examples{
if (interactive()) {
# Launch with built-in PK data set:
guiplot(PK)
}

\dontrun{
# Launch with memory-in data set:
guiplot()

# Launch with memory-in data set, and output plot to user-defined folders:
guiplot(PK,out_dir= Any_directory_you_want_to_export)

}

}
