% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiplot.R
\name{guiplot}
\alias{guiplot}
\title{guiplot}
\usage{
guiplot(...)
}
\arguments{
\item{data}{Matrix or data frame}
}
\value{
code to reproduce chart.
}
\description{
guiplot
}
\examples{
\dontrun{
guiplot(PK)
}
}
